requirements.txt
# Core orchestration and docs
langchain>=0.2.0
langchain-community>=0.2.0

# Claude support in LangChain
langchain-anthropic>=0.1.0

# Vector store
chromadb>=0.5.0

# Local embeddings (no OpenAI required)
sentence-transformers>=2.2.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.7.0

anthropic>=0.25.0

mcp>=0.1.0

nest-asyncio>=1.5.0